Updated to match the example-project-python repo
Updated directory structure and added files to match the example-project-python repo provided.

tests are TODO
