+++ .bumpversion.cfg
@@ -1,0 +1,17 @@
+[bumpversion]
+current_version = 0.0.1
+commit = True
+tag = True
+
+[bumpversion:file:ez_scrape/_version.py]
+search = __version__ = "{current_version}"
+replace = __version__ = "{new_version}"
+
+
+[bumpversion:file:pyproject.toml]
+search = version="{current_version}"
+replace = version="{new_version}"
+
+[bumpversion:file:setup.py]
+search = version="{current_version}"
+replace = version="{new_version}"
