+++ MANIFEST.in
@@ -1,0 +1,20 @@
+graft ez_scrape
+
+include LICENSE
+include README.md
+
+include .bumpversion.cfg
+include pyproject.toml
+include Makefile
+
+prune .github
+exclude .gitignore
+exclude .gitattributes
+
+# Patterns to exclude from any directory
+global-exclude *~
+global-exclude *.pyc
+global-exclude *.pyo
+global-exclude .git
+global-exclude .ipynb_checkpoints
+global-exclude .DS_Store
