+++ ez_scrape/tests/test_all.py
@@ -1,0 +1,12 @@
+from example_project_python import hello, print_hello
+from unittest.mock import patch
+
+
+def test_hello():
+    assert hello() == "Hello, world!"
+
+
+@patch('builtins.print')
+def test_print_hello(mock_print):
+    print_hello()
+    assert mock_print.call_args.args == ("Hello, world!",)
