+++ setup.py
@@ -1,0 +1,4 @@
+from setuptools import setup
+
+
+setup()
